fix(desks): Get rid of RequireJS in desks
