+++ client/test-main.js
@@ -47,7 +47,7 @@ tests.push('superdesk-authoring/multiedit/multiedit');
 tests.push('superdesk-authoring/macros/macros');
 tests.push('superdesk-workspace/content/content');
 tests.push('superdesk-packaging/packaging');
-tests.push('superdesk-desks/module');
+tests.push('superdesk-desks/desks');
 tests.push('superdesk-groups/groups');
 tests.push('superdesk-search/search');
